[Refactoring ⚙️] input과 관련된 validation 기능을 utils 폴더로 분리
